iot/pyproject.toml
[project]
name = "iot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pychromecast>=14.0.7",
    "python-dotenv>=1.1.0",
    "types-requests>=2.32.0.20250515",
]


[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 127
indent-width = 4
exclude = [
    "*/__init__.py",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv/*",
    "local/*",
    "*/prompt.py",
]

[tool.ruff.lint]
ignore = ["B008"]
select = [
    "B",
    "E",
    "F",
    "I",
    "UP",
    "W",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 10

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.11",
]
